requirements.txt
fastapi
uvicorn[standard]
openai-whisper
ffmpeg-python
python-multipart
